Do not generate urls during JSON serialization
The URLs are expire after 5 mins, so there isn't much point considering
it metadata.

Instead, a `download` action has been added which will redirect to the
required S3 bucket.

This ensures the URL is up to date when the download is started. It also
encourages the user not to cache the URL as it isn't downloaded with the
metadata