Redefine the schema to allow a join onto a User
This will be used as implement the `owner` feature on the container.
This replaces the existing `group` indices with a composite of the
`user` and `group`.

They need to be nullable as this will allow the model to enforce a
single owner. A polymorphic key isn't used as they can not enfoce a
foreign reference.