Define the current blobs off the tagged request
Determining the current blobs is a tricky task as it depends on the
scope. It could either refer to the blobs in:
1. current_container, or
2. current_containers

This is toggled by using a dedicated flag.

By using the current_container, the blobs can be filtered by user or
group. This allowed the additional end points in the routes