Replace the multiple end points with a single scope
This removes the duplication between the different end points. Instead
the filtering preformed by each controller can be tweaked by a `scope`.

The `current_user` will always be set if the token is valid. But the
`current_group` can be optional set by setting the scope to :group or
:public.

This in turn tweaks the `current_container(s)`. It also tweaks the
current_blobs list allows for all the blobs for every container to be
listed