Implicitly create tagged containers if required
If a tagged container is referenced directly, it will be implicitly
created if required. This is to take the strain off the admins, so they
don't need to create them manually.

Listing the containers will not create any nor will getting a container
via an `id`.

Tags must be created manually and this is now strictly enforced