Add a dedicated struct to abstract an objects ownership
This removes the ambiguity around what the `containers` method returns.
In the case of a `Group` it returns the containers it owns, but `User`
it returns all the `Containers` they can access.

This behaviour isn't going to be changed, but it introduces an
inconsistency between the objects.

To resolve this, a OwnershipStruct has been created. It is initialized
with the containers a model owns alone. This allows for syntax such as:
`User.first.owns.containers`