Remove the overloading of the `users` method
This causes confusion as it hides which user is attached to the group.
Instead the `readble?` and `writable?` methods handle access to the
public group.

As global admin access is controlled by `cancan`, the models do not need
to worry about this edge case. As such the public group is not
`writable?` to anyone.