Incorporate the offset into the error message
The offset is added to the remaining limit so it gets incorporated into
the error message.

The idea being that the offset represents an old file that is being
edited. Thus after the new file is uploaded the old file is deleted and
thus does not count towards the quota.