Allow the AS::Blob to be nil on a Blob
This is to allow the uploading of the `AS::Blob` to occur after the
`Blob` has been created. This is important as the `Blob` has additional
constraints that are not checked when the `AS::Blob` is created.

However, it isn't a simple case of rollback the changes if the `Blob`
isn't created correctly. This is because the upload has already
occurred, and thus the file needs to be deleted off S3.

For this reason, the upload has been put at the end of the transaction.
The `AS::Blob` is assigned to the blob outside the transaction to
prevent errors from rollback the db. Otherwise there is the risk of
orphans being created in S3