Add a bucket end point that searches by scope/tag
A "bucket" is the same as a Container except it will not search by `id`.
Instead it will use the `:scope` and `:tag` parameters which are
encoded in its path.

Indexing containers are now available over this route. Index uses the
standard `containers` end point with the appropriate query parameters
(either tag or scope, maybe both?)